Cambios 15/10/2019

+++ epsandes/src/main/java/uniandes/isis2304/epsandes/negocio/VOMedico.java
@@ -24,7 +24,7 @@ public interface VOMedico {
 	/**
 	 * @return el numero de registro del medico
 	 */
-	public int numRegMedico();
+	public long numRegMedico();
 	
 
 	/**
